refactor: adjusted the database operation file structure
