+++ app/admin/copyright/page.tsx
@@ -1,4 +1,4 @@
-import { fetchCopyrightList } from '~/server/db/query'
+import { fetchCopyrightList } from '~/server/db/query/copyrights'
 import type { HandleProps } from '~/types/props'
 import React from 'react'
 import CopyrightList from '~/components/admin/copyright/CopyrightList'
