+++ app/admin/page.tsx
@@ -1,4 +1,4 @@
-import { fetchImagesAnalysis } from '~/server/db/query'
+import { fetchImagesAnalysis } from '~/server/db/query/images'
 import CardList from '~/components/admin/dashboard/CardList'
 import type { AnalysisDataProps } from '~/types/props'
