+++ components/layout/DynamicNavbar.tsx
@@ -1,6 +1,6 @@
 import VaulDrawer from '~/components/layout/VaulDrawer'
 import { DropMenu } from '~/components/layout/DropMenu'
-import { fetchAlbumsShow } from '~/server/db/query'
+import { fetchAlbumsShow } from '~/server/db/query/albums'
 import type { AlbumListProps } from '~/types/props'
 import AlbumDrawer from '~/components/layout/AlbumDrawer'
