+++ hono/images.ts
@@ -6,7 +6,7 @@ import {
   updateImage,
   updateImageShow,
   updateImageAlbum
-} from '~/server/db/operate'
+} from '~/server/db/operate/images'
 import { Hono } from 'hono'
 import { HTTPException } from 'hono/http-exception'
